perf: liberally apply #[inline]
In performance profiling, most of these functions aren't being inline
across crate boundaries, which is bad, since most functions are pretty
tiny. As I understand it, in order for non-generic functions to get
inlined across crate boundaries, an #[inline] annotation is required.
Some functions in bstr are generic and so technically don't need them,
but it's easier to audit "put #[inline] on everything" than it is "only
use #[inline] where it's required." Namely, there doesn't seem to be any
harm in using #[inline] if it is otherwise eligible for inlining.

N.B. #[inline] does not *force* a function to get inlined. It is only a
hint.