search: tweak prefilter settings
It turns out that so long as we're skipping a small fixed number of
bytes, then the prefilter is still good to use. I found this when
testing a search of a hash in a large random collection of hashes. This
is good for a 2x speedup without materially impacting existing
benchmarks.

We also bump up the total number of skips that needs to happen before we
consider rendering the prefilter inert. 10 skips isn't a lot and can
cause us to prematurely turn off the prefilter.