api: switch to extension traits
This commit effectively rewrites the entire API of this crate to use
extension traits on `[u8]` and `Vec<u8>`. While the `BStr` and `BString`
types are still present, they are now just dumb wrappers that deref to
`[u8]` and `Vec<u8>`, respectively. Their primary purpose is for
convenient debug impls.

The motivation for this design is laid out in #5.

Closes #5