perf: minimise copying in for_byte_line
Lend out slices of complete strings from the internal buffer, and only
fall back on read_until for fragments at the end.

Because of the increased complexity, we add a number of tests in
addition to the existing doc tests.

This appears to improve performance by about 2x:

    $ critcmp lines01 lines02
    group                       lines01                                lines02
    -----                       -------                                -------
    bstr/for_byte_line/ascii    2.17   403.0±10.00µs  1211.7 MB/sec    1.00    185.4±4.83µs     2.6 GB/sec

Note, however, that this change also uses `find_byte`, which uses the
`memchr` crate. Previously, we were having std scan for the line
terminator with its own implementation of memchr, which isn't as
optimized. So this comparison isn't quite "fair" with respect to
measuring the performance improvement resulting from fewer copies.

Closes #20