impl: improve Debug impl for BStr
Previously, any non-printable Unicode scalar value would get printed
using the '\u{...}' syntax. While not horrible, it's a bit weird to
do this for ASCII characters and especially for the NUL and various
whitespace characters. So this commit tweaks the Debug impl for BStr to
handle those cases specially in a way that's a bit more consistent with
what folks expect.

Fixes #56, Closes #58