impl: support formatting alignment
This adds support for the various'{:5}' syntaxes defined by std::fmt. We
ended up having to re-roll it instead of using std's facilities because
we're working with byte slices and not &str. It would be simple to just
lossily convert a BStr into a String, but this would invalidate the
implementation in no_std contexts.

Fixes #69, Closes #70