impl: fix replacement codepoint handling in Debug impl
Previously, if an actual replacement codepoint was found in UTF-8, it
would print the UTF-8 encoding of the replacement codepoint instead of
the replacement codepoint itself. This is because the replacement
codepoint is used as a sentinel to determine whether invalid UTF-8 was
found. Of course, if the replacement codepoint is encoded itself, then
it's valid UTF-8.

So we fix that by checking whether the raw bytes correspond to a valid
UTF-8 encoding of the replacement codepoint. If so, we print it just
like any other scalar value.

Fixes #72