doc: remove use of the word 'experiment'
While bstr is not yet at 1.0, it is not really an experiment any more.
Its API has and will remain stable for many years to come (with only
very minor breaking changes planned for 1.0, which should happen soon).

In general, the docs used the word "experiment" to refer to two
different things:

1. The idea that "UTF-8 by convention" is more useful in some cases than
   "required valid UTF-8."
2. That dependency trees should be kept to a minimum, i.e., providing a
   cohesive crate rather than a big tree of micro-crates.

(1) seems to be very clearly settled. There are folks using bstr in big
projects precisely because it makes their life a lot easier.

(2) is perhaps still questionable, but it doesn't really make sense to
call it an "experiment" and scare folks away. I think it makes more
sense to just declare "small dependency tree" as a goal that we strive
for.

Closes #99