api: add `Lines::to_bytes()` and `LinesWithTerminator::to_bytes()`
Allowing access to the buffer itself may make some use cases simpler
when access to the original buffer after consuming a few lines is
preferrable.

Without it, one would have to find a way to count the consumed bytes
themselves.

The addition of `to_bytes()` definitely counts as quality of life
improvement for some.