msrv: bump to Rust 1.48
I chose this because it's what appears to be the version of Rust that
is in the current version of Debian stable (which is bullseye).

I don't necessarily plan to always track Debian stable, but in the
absence of any other constraint (like a strong desire to use something
stabilized in a more recent Rust release), it seems like a fine place
to sit.