BREAKING: make splits and find honestly represent their lifetimes
In a few places I must have got lazy when defining the iterator types
and forced haystacks and needles/splitters to always have the same
lifetime. This works in most cases, but #45 shows a case where it breaks
down. To fix it, we just make sure we represent all of our lifetimes in
our types.

Note that #45 was reported against our split types, but we also have
this issue with our find types too. Indeed, our split types are built on
top of our find types, so we just fix everything.

This is a breaking change since we are adding a new lifetime parameter
to several public API types.

Fixes #45