BREAKING: put 'ByteSlice::fields' behind 'unicode' feature
It currently uses 'char::is_whitespace', but this is more of an
implementation detail. While 'char::is_whitespace' is available in
'core', it's plausible that we might use our own data some data. In
particular, 'trim' already uses its own data.

I believe this is the only routine that makes direct use of some kind of
Unicode data that wasn't previously gated behind the 'unicode' feature.

Ref #40