api: add BStr::new
This adds a convenience constructor for building a BStr slice directly
from anything that impls AsRef<[u8]>. This is analogous to, e.g.,
Path::new from std.

While this constructor is redundant, it is often convenient because of
its concreteness. Alternatives rely on generics via trait impls, which
don't always work seamlessly.

Closes #110