doc: enable 'doc_auto_cfg' on docs.rs
This should make it easier to see which APIs need to have certain
features enabled.

This unfortunately is not perfect, which is probably why it's unstable.
For example, it doesn't show that types like 'Graphemes' are only
available with the 'unicode' feature enabled, presumably because the
re-exports trip up rustdoc somehow. But, the 'ByteSlice::graphemes'
routine is correctly marked as requiring the 'unicode' feature, which is
something, I guess?

Let's try it out and see if folks have feedback.