safety: don't use integer-to-pointer casts
Miri warns about this, so we fix it. I have no particular attachment to
integer-to-pointer casts, and it seems like the tide might be shifting
against them. Since we don't need an integer-to-pointer cast here, we
drop it. I believe it was written the way it was (from the 'memchr
crate, which itself should be fixed) because it made more intuitive
sense to me. Alas...