BREAKING: api: tweak wording of OsStr/Path conversions
This slightly tweaks the documentation of the various OsStr/Path
conversion routines to cover the possible future where `OsStr::as_bytes`
(and probably also `OsStr::from_bytes`) exists. Namely, the point of
these routines was always to get at the underlying byte representation
of OS strings, but doing so without building a WTF-8 decoder/encoder. So
if the raw bytes do wind up getting exposed in a platform independent
way, then we'll switch all of these routines to use the raw bytes and
they'll all become, effectively, no-ops.

But for now, we still continue to choose to sacrifice a little
correctness on Windows in exchange for a lot of simplicity.

Closes #115