msrv: bump to Rust 1.60
And switch to Rust 2011 and 'resolver = 2'.

The idea here is to overhaul bstr features to make use of the new
resolver in Cargo. But this requires Rust 1.60.

This does make the MSRV a bit newer than I would hope, but it's at least
a few releases old at this point. And if I don't do this now, then I
either have to accept a sub-optimal feature setup for 1.0 forever, or
delay 1.0 until Rust 1.60 is "old enough."

Instead, we can just bump to Rust 1.60 now. For people too squeamish to
move to such a new Rust, they can stick to bstr 0.2.0 until they're
ready to move on.

Ref: https://github.com/BurntSushi/bstr/issues/40#issuecomment-1180723747