BREAKING: cargo: simplify crate features
This gets rid of the 'serde1-{std,alloc,core}' features and just
replaces it with a single proper 'serde' feature. We make it work as one
would expect with the other crate features by using the new 'dep:' and
'pkg?' Cargo.toml feature syntax.

This is a breaking change because the 'serde1' features no longer exist.
Instead, you just need to enable the 'serde' feature and it should
automatically do the right thing depending on whether you also have
'std', 'alloc' or neither set.

Ref: https://doc.rust-lang.org/cargo/reference/features.html
Ref: https://github.com/BurntSushi/bstr/issues/40#issuecomment-1180723747