ci: remove 'alloc,unicode' test
It doesn't compile any more, by design. The issue here is that enabling
'alloc,unicode' makes it look like you're getting rid of std, but it
would silently bring in a dependency on 'std' via 'once_cell'. So as
explained in the previous commit, we require 'std' to be explicitly
enabled. That way, in the future, we can drop 'once_cell' and all re-add
the 'alloc,unicode' test.