bench: remove lock file
The 'bench' crate was added to the bstr workspace, so now there is no
lock file. Which seems kind of bad to be honest. I don't usually commit
lock files for libraries, but do for their benchmarks.

Let's see how far we get with this and if any problems appear.

Fixes #131