Make `Sealed` super trait impossible to export
Drive by refactor after revisiting #133.

The `Sealed` trait in `crate::ext_slice` and `crate::ext_vec` is marked
`pub` and is reachable from the crate root, which means there is a
(small) risk the trait may be accidentally exported as public API.

The sealed trait trick works so long as the trait is marked `pub` even
if the trait is not public to the whole crate.

This commit wraps `pub trait Sealed` in a private module, which ensures
the sealed trait in `crate::ext_slice` may only be used for implementing
`ByteSlice` (and similarly for `crate::ext_vec` and `ByteVec`).

The new sealed trait is defined and used like this:

```rust
mod private {
    pub trait Sealed {}
}

impl private::Sealed for &[u8] {}

trait ByteSlice: private::Sealed {
    // ...
}
```

This approach is discussed at
https://rust-lang.github.io/api-guidelines/future-proofing.html.

A short addition to the `ByteSlice` and `ByteVec` traits indicate that
these traits are sealed and not meant to be implemented outside of
`bstr`.