impl: make `Sealed` super trait impossible to export
The `Sealed` trait in `crate::ext_slice` and `crate::ext_vec` is marked
`pub` and is reachable from the crate root, which means there is a
(small) risk the trait may be accidentally exported as public API.

The sealed trait trick works so long as the trait is marked `pub` even
if the trait is not public to the whole crate.

This commit wraps `pub trait Sealed` in a private module, which ensures
the sealed trait in `crate::ext_slice` may only be used for implementing
`ByteSlice` (and similarly for `crate::ext_vec` and `ByteVec`).

This approach is discussed at
https://rust-lang.github.io/api-guidelines/future-proofing.html.

Note that this doesn't make any public API changes. The `Sealed` trait
was not previously exposed. This commit is just about making it harder
to *accidentally* expose it in the future.

PR #144