api: add ByteSlice::escape_bytes and ByteVec::unescape_bytes
I grew weary of re-implementing basically these exact routines all of
the time. In my work, it comes up surprisingly often that I want to be
able to accept any kind of input---including arbitrary bytes---but
either don't want to force the user to figure out how to type in invalid
UTF-8, or are otherwise in a context where invalid UTF-8 is forbidden
(like a TOML file).

The two functions introduced here are duals of one another. The unescape
routine knows how to get back the raw bytes, where as the escape routine
knows how to take raw bytes and make them "human readable."

One potential downside of this mechanism is that it doesn't support any
kind of Unicode escapes. It was partially for that reason that I called
this new API 'escape_bytes', where 'bytes' puts a focus on what is
actually being done here. Plus, Unicode escpaes open their own can of
worms (as usual), with questions about what to do in the case of an
invalid codepoint number and which codepoints to escape in the first
place.

Both of those questions probably have good answers, but I just didn't
want to go down that path for now.