unicode/fsm: switch to regex-automata 0.3
This changes the Unicode generation script to use 'regex-cli' instead of
'ucd-generate'. Then I generated the new DFAs and fixed the fallout in
the code. The new APIs are a little more clunky, but such is life.

Amazingly, everything Just Worked. And the generation script now takes 5
seconds instead of ~40 seconds(!).